Renaming RNAdam to rice.
