--- bin/RNAdam-submit
@@ -1,61 +1,0 @@
-#!/usr/bin/env bash
-#
-# Licensed to Big Data Genomics (BDG) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The BDG licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# Figure out where RNADAM is installed
-RNADAM_REPO="$(cd `dirname $0`/..; pwd)"
-
-CLASSPATH=$("$RNADAM_REPO"/bin/compute-RNAdam-classpath.sh)
-RNADAM_JARS=$("$RNADAM_REPO"/bin/compute-RNAdam-jars.sh)
-
-# Find the RNADAM CLI jar
-num_versions=$(ls "$RNADAM_REPO"/RNAdam-cli/target/appassembler/repo/org/bdgenomics/RNAdam/RNAdam-cli | wc -l)
-if [ "$num_versions" -eq "0" ]; then
-  echo "Failed to find RNAdam-cli jar in $RNADAM_REPO/RNAdam-cli/target/appassembler/repo/org/bdgenomics/RNAdam/RNAdam-cli"
-  echo "You need to build RNAdam before running this program."
-  exit 1
-fi
-if [ "$num_versions" -gt "1" ]; then
-  versions_list=$(ls "$RNADAM_REPO"/RNAdam-cli/target/appassembler/repo/org/bdgenomics/RNAdam/RNAdam-cli)
-  echo "Found multiple RNAdam CLI versions in $RNADAM_REPO/RNAdam-cli/target/appassembler/repo/org/bdgenomics/RNAdam/RNAdam-cli:"
-  echo "$versions_list"
-  echo "Please remove all but one."
-  exit 1
-fi
-RNADAM_CLI_JAR=$(ls "$RNADAM_REPO"/RNAdam-cli/target/appassembler/repo/org/bdgenomics/RNAdam/RNAdam-cli/*/RNAdam-cli-*.jar)
-
-if [ -z "$SPARK_HOME" ]; then
-  echo "Attempting to use 'spark-submit' on default path; you might need to set SPARK_HOME"
-  SPARK_SUBMIT=spark-submit
-else
-  SPARK_SUBMIT="$SPARK_HOME"/bin/spark-submit
-fi
-
-# submit the job to Spark
-"$SPARK_SUBMIT" \
-  --class org.bdgenomics.RNAdam.cli.RNAdamMain \
-  --conf spark.serializer=org.apache.spark.serializer.KryoSerializer \
-  --conf spark.kryo.registrator=org.bdgenomics.adam.serialization.ADAMKryoRegistrator \
-  --conf spark.kryoserializer.buffer.mb=4 \
-  --conf spark.kryo.referenceTracking=true \
-  --conf spark.executor.memory=${RNADAM_EXECUTOR_MEMORY:-4g} \
-  ${RNADAM_OPTS:- } \
-  --driver-memory ${RNADAM_DRIVER_MEMORY:-4g} \
-  --jars "$RNADAM_JARS" \
-  "$RNADAM_CLI_JAR" \
-  "$@"
