+++ bin/rice-submit
@@ -1,0 +1,61 @@
+#!/usr/bin/env bash
+#
+# Licensed to Big Data Genomics (BDG) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The BDG licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+# Figure out where RICE is installed
+RICE_REPO="$(cd `dirname $0`/..; pwd)"
+
+CLASSPATH=$("$RICE_REPO"/bin/compute-rice-classpath.sh)
+RICE_JARS=$("$RICE_REPO"/bin/compute-rice-jars.sh)
+
+# Find the RICE CLI jar
+num_versions=$(ls "$RICE_REPO"/rice-cli/target/appassembler/repo/org/bdgenomics/rice/rice-cli | wc -l)
+if [ "$num_versions" -eq "0" ]; then
+  echo "Failed to find rice-cli jar in $RICE_REPO/rice-cli/target/appassembler/repo/org/bdgenomics/rice/rice-cli"
+  echo "You need to build rice before running this program."
+  exit 1
+fi
+if [ "$num_versions" -gt "1" ]; then
+  versions_list=$(ls "$RICE_REPO"/rice-cli/target/appassembler/repo/org/bdgenomics/rice/rice-cli)
+  echo "Found multiple rice CLI versions in $RICE_REPO/rice-cli/target/appassembler/repo/org/bdgenomics/rice/rice-cli:"
+  echo "$versions_list"
+  echo "Please remove all but one."
+  exit 1
+fi
+RICE_CLI_JAR=$(ls "$RICE_REPO"/rice-cli/target/appassembler/repo/org/bdgenomics/rice/rice-cli/*/rice-cli-*.jar)
+
+if [ -z "$SPARK_HOME" ]; then
+  echo "Attempting to use 'spark-submit' on default path; you might need to set SPARK_HOME"
+  SPARK_SUBMIT=spark-submit
+else
+  SPARK_SUBMIT="$SPARK_HOME"/bin/spark-submit
+fi
+
+# submit the job to Spark
+"$SPARK_SUBMIT" \
+  --class org.bdgenomics.rice.cli.RiceMain \
+  --conf spark.serializer=org.apache.spark.serializer.KryoSerializer \
+  --conf spark.kryo.registrator=org.bdgenomics.adam.serialization.ADAMKryoRegistrator \
+  --conf spark.kryoserializer.buffer.mb=4 \
+  --conf spark.kryo.referenceTracking=true \
+  --conf spark.executor.memory=${RICE_EXECUTOR_MEMORY:-4g} \
+  ${RICE_OPTS:- } \
+  --driver-memory ${RICE_DRIVER_MEMORY:-4g} \
+  --jars "$RICE_JARS" \
+  "$RICE_CLI_JAR" \
+  "$@"
